+++ public/index.html
@@ -9,7 +9,7 @@
             name="description"
             content="Web site created using create-react-app"
         />
-        <title>React App</title>
+        <title>Kasa</title>
     </head>
     <body>
         <div id="root"></div>
