+++ src/styles/base/_typography.scss
@@ -1,4 +1,5 @@
 @import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@400;700&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@500&display=swap");
 
 body,
 * {
@@ -52,14 +53,15 @@ h4 {
 }
 
 li {
-    font-size: 0.7rem;
+    font-size: 0.8rem;
     @include desktop-only {
-        font-size: 0.8rem;
+        font-size: 1rem;
     }
 }
 
 a {
     font-size: 0.8rem;
+    color: $color-piment;
     @include desktop-only {
         font-size: 1.3rem;
     }
