+++ src/styles/layout/_banner.scss
@@ -1,0 +1,39 @@
+.banner {
+    background-color: black;
+    border-radius: 25px;
+    &__homeimg {
+        background: url("../../assets/backheader.png") center/cover;
+        border-radius: 25px;
+        height: 111px;
+        width: 100%;
+        @include desktop-only {
+            height: 223px;
+            justify-content: center;
+        }
+    }
+    &__title {
+        color: $color-white;
+        background-color: rgb(000 000 000 / 0.3);
+        border-radius: 25px;
+        height: 111px;
+        display: flex;
+        align-items: center;
+        justify-content: flex-start;
+        padding: 20px;
+        @include desktop-only {
+            height: 223px;
+            justify-content: center;
+        }
+    }
+    &__aboutimg {
+        background: url("../../assets/aboutbanner.png") center/cover;
+        border-radius: 25px;
+        height: 223px;
+        width: 100%;
+        opacity: 80%;
+        @include desktop-only {
+            height: 223px;
+            justify-content: center;
+        }
+    }
+}
