+++ src/styles/layout/_collapses.scss
@@ -18,6 +18,9 @@
         border-radius: 5px;
         padding: 20px 10px 10px 10px;
         color: $color-piment;
+        @include mobile-only {
+            padding-bottom: 60px;
+        }
 
         animation: collapsein $menu-dur $menu-delay ease-in-out both;
     }
