+++ src/styles/layout/_nav.scss
@@ -16,11 +16,12 @@
         margin-left: auto;
         & li {
             color: $color-piment;
-            margin-left: 20px;
+            margin-left: 50px;
             font-size: 1.3rem;
             @include mobile-only {
                 text-transform: uppercase;
                 font-size: 0.8rem;
+                margin-left: 30px;
             }
         }
         &__active {
