+++ src/styles/utils/_variables.scss
@@ -1,10 +1,9 @@
-
 // Fonts
 $font-1: "Montserrat", sans-serif;
 
 //Colors
-$color-gray: #F6F6F6;
-$color-piment: #FF6060;
+$color-gray: #f6f6f6;
+$color-piment: #ff6060;
 $color-darkgray: #666666;
 $color-white: white;
 $color-black: black;
@@ -13,6 +12,5 @@ $color-black: black;
 $layout-mobile-padding: 20px;
 $layout-desktop-padding: 40px;
 
-
-$menu-dur:200ms;
-$menu-delay:150ms;
+$menu-dur: 200ms;
+$menu-delay: 150ms;
